refactor(src/): #1 - Refactor App and Put Scaffolding In Place
Broke out the app a bit to make it easier to work on parallel. Added a bit of
scaffolding to get us started. Nothing set in stone. Just some basics that
can be moved around. The component folder should allow us to work on
components with out stepping on toes.

Converted CSS to CSS Modules (supported out of the box with CRA)